+++ packages/dart_firebase/lib/src/types/firestore/reference.dart
@@ -1,0 +1,6 @@
+part of dart_firebase;
+
+abstract class FirestoreReference {
+  FirestoreClient get client;
+  List<String> get pathComponents;
+}