+++ packages/dart_firebase/lib/src/types/object.dart
@@ -1,0 +1,6 @@
+part of dart_firebase;
+
+abstract class FirestoreObject {
+  FirestoreClient get client;
+  Map<String, dynamic> get json;
+}